добавлены константы продолжит дней, evt.key
