+++ src/utils/date.js
@@ -1,8 +1,10 @@
-import {getRandomInteger} from "../mock/utils.js";
 
 const MINUTE = 60 * 1000;
 const HOUR = 60 * MINUTE;
 const DAY = 24 * HOUR;
+const HOURS_PER_DAY = 24;
+const MINUTES_PER_HOUR = 60;
+
 
 const convertDate = () => {
   const date = new Date();
@@ -22,35 +24,18 @@ const convertDateNumbers = (value) => String(value).padStart(2, `0`);
 const formatDuration = (duration) => {
 
   const durationDays = Math.floor(duration / DAY);
-  const durationDaysString = durationDays > 0 ? `${convertDateNumbers(durationDays)}D` : ``;
+  const durationDaysResult = durationDays > 0 ? `${convertDateNumbers(durationDays)}D` : ``;
 
-  const durationHours = Math.floor(duration / HOUR % 24);
-  const durationHoursString = durationHours > 0 ? `${convertDateNumbers(durationHours)}H` : ``;
+  const durationHours = Math.floor(duration / HOUR % HOURS_PER_DAY);
+  const durationHoursResult = durationDays > 0 || durationHours > 0 ? `${convertDateNumbers(durationHours)}H` : ``;
 
-  const durationMinutes = Math.floor(duration / MINUTE % 60);
-  const durationMinutesString = durationMinutes > 0 ? `${convertDateNumbers(durationMinutes)}M` : ``;
+  const durationMinutes = Math.floor(duration / MINUTE % MINUTES_PER_HOUR);
+  const durationMinutesResult = durationMinutes > 0 ? `${convertDateNumbers(durationMinutes)}M` : ``;
 
-  return `${durationDaysString} ${durationHoursString} ${durationMinutesString}`;
+  return `${durationDaysResult} ${durationHoursResult} ${durationMinutesResult}`;
 };
 
 const getFirstUpperCase = (element) => element[0].toUpperCase() + element.slice(1);
-
-const getStartDate = () => {
-  const maxDayRange = 4;
-  const dayRange = getRandomInteger(1, maxDayRange);
-  const currentDate = new Date();
-  currentDate.setHours(getRandomInteger(0, 23), getRandomInteger(0, 59));
-  currentDate.setDate(currentDate.getDate() + dayRange);
-  return new Date(currentDate);
-};
-
-const getEndDate = (date) => {
-  const startDate = new Date(date);
-  const min = new Date(startDate).getTime();
-  const max = new Date(startDate.setHours(24, 59, 59, 999)).getTime();
-  const endDate = new Date(getRandomInteger(min, max));
-  return new Date(endDate);
-};
 
 const convertDateToDay = (date) => date.toISOString().slice(0, 10);
 
@@ -70,4 +55,4 @@ const groupCardsByDay = (sortedCards) => {
   return sortedCards.reduce(reduceCardByDay, {});
 };
 
-export {MINUTE, HOUR, DAY, getTimeFormat, getDayFormat, convertDate, formatDuration, getFirstUpperCase, getStartDate, getEndDate, groupCardsByDay};
+export {getTimeFormat, getDayFormat, convertDate, formatDuration, getFirstUpperCase, groupCardsByDay};
